# backend/docs/swagger.yaml
basePath: /v1
definitions:
  entities.Author:
    properties:
      icon_url:
        type: string
      login:
        type: string
    type: object
  entities.Comment:
    properties:
      author:
        $ref: '#/definitions/entities.Author'
      created_at:
        type: string
      id:
        type: integer
      text:
        maxLength: 250
        minLength: 1
        type: string
      updated_at:
        type: string
    required:
    - text
    type: object
  entities.CommentCreate:
    properties:
      text:
        maxLength: 250
        minLength: 1
        type: string
    required:
    - text
    type: object
  entities.CommentDelete:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  entities.CommentUpdate:
    properties:
      id:
        type: integer
      text:
        maxLength: 250
        minLength: 1
        type: string
    required:
    - id
    - text
    type: object
  entities.FullRecipe:
    properties:
      author:
        $ref: '#/definitions/entities.Author'
      comments:
        items:
          $ref: '#/definitions/entities.Comment'
        type: array
      is_liked:
        type: boolean
      likes_count:
        type: integer
      recipe:
        $ref: '#/definitions/entities.Recipe'
    type: object
...
  entities.Recipe:
    properties:
      about:
        maxLength: 2500
        type: string
      complexitiy:
        enum:
        - 1
        - 2
        - 3
        maximum: 3
        minimum: 1
        type: integer
      created_at:
        type: string
      id:
        type: integer
      ingridients:
        maxLength: 1500
        type: string
      need_time:
        type: string
      photos_urls:
        type: string
      title:
        maxLength: 50
        minLength: 3
        type: string
      updated_at:
        type: string
    required:
    - about
    - complexitiy
    - ingridients
    - need_time
    - title
    type: object
...
  entities.RecipeInfo:
    properties:
      info:
        $ref: '#/definitions/entities.FullRecipe'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: RestAPI for recipe site
  title: RecipeSite
  version: "1.0"
paths:
...
  /recipe/{id}:
    get:
      description: Get recipe
      operationId: get recipe
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RecipeInfo'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get recipe
      tags:
      - recipe
  /recipe/{id}/comment:
    delete:
      consumes:
      - application/json
      description: Delete comment
      operationId: delete comment
      parameters:
      - description: Comment params
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/entities.CommentDelete'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Delete comment
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Create comment
      operationId: create comment
      parameters:
      - description: Comment params
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/entities.CommentCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Create comment
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Update comment
      operationId: update comment
      parameters:
      - description: Comment params
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/entities.CommentUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Update comment
      tags:
      - comments
  /recipe/{id}/like:
    post:
      description: Like recipe
      operationId: like
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Like
      tags:
      - likes
  /recipe/{id}/unlike:
    post:
      description: Unlike recipe
      operationId: unlike
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Unlike
      tags:
      - likes
...
swagger: "2.0"